# Bugfix ChainedVector constructor performance